FlaskServer/requirements.txt
Flask==2.3.5
requests==2.32.4
pandas==2.1.0
numpy==1.27.0
scikit-learn==1.5.0
python-dotenv==1.0.1
